+++ src/pages/api/unit.js
@@ -1,7 +1,13 @@
 import db from '../../lib/db';
 
 export default async function handler(req, res) {
+    const apiKey = req.headers['apikey'];
+    const envApiKey = process.env.NEXT_PUBLIC_API_KEY;
+    
     if (req.method == 'GET') {
+        if (apiKey !== envApiKey) {
+            return res.status(403).json({ error: 'Akses ditolak !' });
+        }
         try {
             const [rows] = await db.query('SELECT * FROM tbl_unit');
             res.status(200).json(rows);
@@ -9,6 +15,9 @@ export default async function handler(req, res) {
             res.status(500).json({ error: 'Gagal fetching data' });
         }
     }else if (req.method == 'POST') {
+        if (apiKey !== envApiKey) {
+            return res.status(403).json({ error: 'Akses ditolak !' });
+        }
         const { textUnit } = req.body;
         try {
             await db.query('INSERT INTO tbl_unit (unit) VALUES (?)', [textUnit]);
@@ -17,6 +26,9 @@ export default async function handler(req, res) {
             res.status(500).json({ error: 'Gagal menambahkan unit' });
         }
     }else if (req.method == 'PUT') {
+        if (apiKey !== envApiKey) {
+            return res.status(403).json({ error: 'Akses ditolak !' });
+        }
         const { updatedId, updatedUnit} = req.body;
         try {
             await db.query('UPDATE tbl_unit SET unit = ? WHERE tbl_unit.id = ?', [updatedUnit, updatedId]);
@@ -25,6 +37,9 @@ export default async function handler(req, res) {
             res.status(500).json({ error: 'Gagal memperbarui unit' });
         }
     }else if (req.method == 'DELETE') {
+        if (apiKey !== envApiKey) {
+            return res.status(403).json({ error: 'Akses ditolak !' });
+        }
         const { deletedId } = req.body;
         try {
             await db.query('DELETE FROM tbl_unit WHERE id = ?', [deletedId]);
